Update Document Intelligence models and tests; change Width property to double, add JsonIgnore to Content, and update test file paths

+++ src/HPHA.UiPath.Core/Azure/DocumentIntelligence/Page.cs
@@ -11,7 +11,7 @@ namespace HPHA.UiPath.Core.Azure.DocumentIntelligence
         public double? Angle { get; set; }
 
         [JsonPropertyName("width")]
-        public int? Width { get; set; }
+        public double? Width { get; set; }
 
         [JsonPropertyName("height")]
         public double? Height { get; set; }
